changing dtype to double

+++ include/complex.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 struct Complex{
-    float mReal;
-    float mImag;
+    double mReal;
+    double mImag;
 
     Complex operator* (const Complex& aOther);
     Complex& operator+= (const Complex& aOther);
@@ -10,7 +10,7 @@ struct Complex{
     //modify the above to return a newly constructed object
 
     // need a constrcutor that defaults to 0
-    Complex(float aReal, float aImag);
+    Complex(double aReal, double aImag);
 
 
     Complex(const Complex& aOther);
@@ -18,7 +18,7 @@ struct Complex{
 
     void display() const;
 
-    float absolute() const;
+    double absolute() const;
   };
